+++ frb_codegen/src/library/codegen/ir/field.rs
@@ -20,6 +20,16 @@ pub struct IrFieldSettings {
 }
 }
 
+#[derive(Debug, Clone, Hash, Eq, PartialEq, serde::Serialize, strum_macros::Display)]
+pub enum OwnershipMode {
+    /// "T"
+    Owned,
+    /// "&T"
+    Ref,
+    /// "&mut T"
+    RefMut,
+}
+
 impl IrField {
     #[inline]
     pub fn is_optional(&self) -> bool {
