+++ frb_codegen/src/library/codegen/parser/type_parser/rust_auto_opaque.rs
@@ -1,7 +1,8 @@
 use crate::codegen::ir::namespace::Namespace;
 use crate::codegen::ir::ty::rust_auto_opaque::{
-    IrRustAutoOpaqueRaw, IrTypeRustAutoOpaque, OwnershipMode,
+    IrRustAutoOpaqueRaw, IrTypeRustAutoOpaque
 };
+use crate::codegen::ir::field::OwnershipMode;
 use crate::codegen::ir::ty::rust_opaque::{
     IrRustOpaqueInner, IrTypeRustOpaque, RustOpaqueCodecMode,
 };
